:recycle: Refactor code to fit latest pypi structure
- Use pyproject.toml file instead of setup.py
- Add MIT License file
- Fix __main__.py to call `run_nxpprog` and not `main` which works
  without causing an exception
- Bump version to 1.0.3
