--- requirements.txt
@@ -1,4 +1,0 @@
-intelhex==2.1
-pyserial==3.4
-click==7.1.2
-rich==12.6.0
