+++ src/nxpprog/__main__.py
@@ -1,0 +1,4 @@
+from nxpprog import nxpprog
+import sys
+
+nxpprog.run_nxpprog(sys.argv)
